# Conflicts:
#	app/Portal.Consultoras.Common/Constantes.cs
#	app/Portal.Consultoras.Entities/BEMisPedidos.cs
#	app/Portal.Consultoras.Web/Scripts/PortalConsultoras/MisPedidos/ClienteOnline.js
#	app/Portal.Consultoras.Web/Views/MisPedidos/Index.cshtml

+++ app/Portal.Consultoras.ServiceContracts/IUsuarioService.cs
@@ -164,7 +164,7 @@ namespace Portal.Consultoras.ServiceContracts
         IList<BEMisPedidos> GetMisPedidosConsultoraOnline(int PaisID, long ConsultoraId, int Campania);
 
         [OperationContract]
-        IList<BEMisPedidosDetalle> GetMisPedidosDetalleConsultoraOnline(int PaisID, int PedidoID);
+        IList<BEMisPedidosDetalle> GetMisPedidosDetalleConsultoraOnline(int PaisID, long PedidoID);
 
         [OperationContract]
         IList<BEMisPedidos> GetMisPedidosClienteOnline(int paisID, long consultoraId, int campania);
